requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1

pytest==8.3.2